requirements.txt
amqp==2.1.3
billiard==3.5.0.3
celery==4.1.0
Django==1.11
django-filter==1.1.0
django-viewflow==1.1.0
pytz==2017.3
# redis==2.10.6
